openai/requirements.txt
Flask
flask-cors
python-telegram-bot==13.15
python-dotenv
google-generativeai
pandas
openpyxl
urllib3==1.26.6
six
